@@ -26,14 +26,14 @@
     "@types/mathjs": "^9.4.2",
     "@types/node": "^12.20.37",
     "@types/pako": "^1.0.3",
-    "@types/react": "^17.0.37",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "color": "^4.2.3",
     "install": "^0.13.0",
     "mathjs": "^10.1.0",
     "pako": "^2.0.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-scripts": "4.0.3",
     "three": "^0.139.2",
     "ts-jest": "^27.1.4",
